+++ wgpu-core/src/binding_model.rs
@@ -889,6 +889,20 @@ pub enum ImmediateUploadError {
         size: u32,
         immediate_size: u32,
     },
+    #[error("Start index {start_index} overruns the value data range with {data_size} element(s)")]
+    ValueStartIndexOverrun { start_index: u32, data_size: usize },
+    #[error(
+        "Start index {} + count of {} overruns the value data range \
+        with {} element(s)",
+        start_index,
+        count,
+        data_size
+    )]
+    ValueEndIndexOverrun {
+        start_index: u32,
+        count: u32,
+        data_size: usize,
+    },
 }
 
 impl WebGpuError for ImmediateUploadError {