+++ wgpu-core/src/command/pass.rs
@@ -240,8 +240,30 @@ where
 
     pipeline_layout.validate_immediates_ranges(offset, size_bytes)?;
 
-    let values_end_offset = (values_offset + size_bytes / wgt::IMMEDIATE_DATA_ALIGNMENT) as usize;
-    let data_slice = &immediates_data[(values_offset as usize)..values_end_offset];
+    let values_offset_usize = values_offset as usize;
+    if values_offset_usize > immediates_data.len() {
+        return Err(ImmediateUploadError::ValueStartIndexOverrun {
+            start_index: values_offset,
+            data_size: immediates_data.len(),
+        }
+        .into());
+    }
+
+    // NOTE: The `validate_immediates_ranges` call above validates `size_bytes` is aligned.
+    let size_immediate_elements = size_bytes / wgt::IMMEDIATE_DATA_ALIGNMENT;
+    let size_immediate_elements_usize = size_immediate_elements as usize;
+    if size_immediate_elements_usize > immediates_data.len() - values_offset_usize {
+        return Err(ImmediateUploadError::ValueEndIndexOverrun {
+            start_index: values_offset,
+            count: size_immediate_elements,
+            data_size: immediates_data.len(),
+        }
+        .into());
+    }
+
+    // NOTE: These additions are will not overflow, because we've validated the range above.
+    let values_end_offset = values_offset_usize + size_immediate_elements_usize;
+    let data_slice = &immediates_data[(values_offset_usize)..values_end_offset];
 
     f(data_slice);